backend/requirements.txt
azure-ai-formrecognizer
azure-identity
azure-storage-blob
dj-database-url
Django
django-allauth
django-appconf
django-bootstrap4
django-celery-beat
django-ckeditor
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-filter
django-formset-js-improved
django-static-fontawesome
django-storages
gunicorn
h11
Pillow
python-dotenv
h2
hpack
hstspreload
whitenoise
psycopg2-binary
opencv-contrib-python
tensorflow[and-cuda]
torch
torchvision
torchaudio
transformers
tk
numpy
pandas
opencv-python
PyMuPDF
cloud-tpu-client
